implement ziospecabstract
